+++ src/Forms/Password.js
@@ -1,23 +1,16 @@
 import React, { useState, useEffect } from "react";
 import { Link, withRouter } from "react-router-dom";
 import { connect } from "react-redux";
-import Spinner from "./UI/Spinner";
+import Spinner from "../UI/Spinner";
 
-function Login(props) {
-  const [user, setUser] = useState();
-  const { history } = props;
-
-  let getEmail = (event) => {
-    setUser({ ...user, email: event.target.value });
-  };
-  let getPass = (event) => {
-    setUser({ ...user, password: event.target.value });
-  };
+function Password(props) {
+  const [email, setEmail] = useState("");
+  const { loading, dispatch } = props;
   useEffect(() => {
     if (localStorage.token) {
       props.history.push("/");
     }
-  }, [props.isLogin, props.history]);
+  }, [props.history]);
 
   const [errorMessage, seterrorMessage] = useState({});
   const validate = (elements) => {
@@ -32,9 +25,6 @@ function Login(props) {
         errors.email = "Plaese enter valid email";
       }
     }
-    if (!elements.password.value) {
-      errors.password = "Plaese enter password";
-    }
     var errorKeys = Object.keys(errors);
     if (errorKeys.length > 0) {
       return errors;
@@ -45,58 +35,44 @@ function Login(props) {
 
   const submit = (event) => {
     event.preventDefault();
-    var form = document.getElementById("loginform");
+    var form = document.getElementById("forgotform");
     var errors = validate(form.elements);
 
     if (errors) {
       seterrorMessage(errors);
     } else {
       seterrorMessage({});
-
-      props.dispatch({
-        type: "LOGIN",
-        payload: user,
-        history: history,
+      dispatch({
+        type: "FORGOT_PASSWORD_INIT",
+        payload: { email: email },
       });
     }
   };
   return (
     <>
-      {props.loading ? (
+      {loading ? (
         <Spinner />
       ) : (
-        <form id="loginform" className="custom-form">
-          <h2 style={{ textAlign: "center" }}>Login</h2>
+        <form id="forgotform" className="custom-form">
+          <h2 style={{ textAlign: "center" }}>Forgot Password</h2>
           <div className="form-group">
             <label>Email</label>
             <input
               type="email"
               name="email"
               className="form-control"
-              onChange={getEmail}
+              value={email}
+              onChange={(event) => setEmail(event.target.value)}
             />
             <span style={{ color: "red" }}>{errorMessage?.email}</span>
           </div>
-          <div className="form-group">
-            <label>Password</label>
-            <input
-              type="password"
-              name="password"
-              className="form-control"
-              onChange={getPass}
-            />
-            <span style={{ color: "red" }}>{errorMessage?.password}</span>
-          </div>
           <button className="btn btn-primary" onClick={submit}>
-            Login
+            Submit
           </button>
           <br></br>
           <br></br>
           <div style={{ float: "left" }}>
-            <Link to="/signup">New User? Click here</Link>
-          </div>
-          <div style={{ float: "right" }}>
-            <Link to="/forgot-password">Forgot Password</Link>
+            <Link to="/login">Login ? CLick here</Link>
           </div>
           <br></br>
         </form>
@@ -108,7 +84,5 @@ function Login(props) {
 export default connect(function (state, props) {
   return {
     loading: state?.isFetching,
-    isLogin: state?.isLogin,
-    isError: state?.isLoginError,
   };
-})(withRouter(Login));
+})(withRouter(Password));
